수정: List.js , listStore mobx 적용시 코드 변경외
