+++ packages/frontend/package.json
@@ -6,7 +6,6 @@
 		"@babel/core": "7.2.2",
 		"@material-ui/core": "^3.9.3",
 		"@svgr/webpack": "4.1.0",
-		"axios": "^0.19.0",
 		"babel-core": "7.0.0-bridge.0",
 		"babel-eslint": "9.0.0",
 		"babel-jest": "23.6.0",
