+++ packages/frontend/src/index.js
@@ -5,6 +5,7 @@ import App from "./Components/App";
 import * as serviceWorker from "./serviceWorker";
 import RootStore from "./Stores";
 
+// 스토어 인스턴스 생성
 const root = new RootStore();
 
 ReactDOM.render(<Provider {...root}><App /></Provider>, document.getElementById("root"));
