+++ yu_report/yu_190617_MobX.md
@@ -30,20 +30,35 @@
 
 </pre>
 
+* Store폴더 생성후 관련 파일 작성
 
 * mobx를 사용할 때에는 observable를 사용하여 state를 관리
-
-
-* Store폴더 생성후 관련 파일 작성
-* 
-* 파일에 state , action 로직 위치 이동
+<pre>
+	@observable items = [];
+	@observable numOfRows = 72;
+	@observable pageNo = 1;
+	@observable scrolling = false;
+	@observable hasMore = true;
+	@observable isLoading = false;
+	@observable error = false;
+</pre>
+ 
+*  state , action 로직 위치 이동
   
-* setState는 불필요한 렌더링을 만듦
-		@action 은 setSate
-		@computed는 this.state 
+   * setState는 불필요한 렌더링을 만듦
+   * @action으로 state를 동기적 방식으로 변경할수 있다.
 
-### 고민..
- 1. hook으로 리팩토링을 한 후  mobx와 적용하는게 나을까? 라이프사이클상태에서 진행해야 할까?
- 2. 후자라면 mobx와 React lifecycle은 어떻게 사용해야 할까?
- 3. async 함수는 어떻게 사용하나?
+* Provider로 프로젝트에 스토어 적용
+	* App 컴포넌트를 Provider컴포넌트로 살포시 감싸준다.
+
+* 스토어를 분리시킨 해당 컴포넌트에 @observer를 사용한다.
+
+* 그 위에 @inject('listStore')를 넣어주면  listStore를 props로 전달받을 수 있게 된다.
+
+* async, await은 
+
+
+
+
+
